Доделать регистрацию
