+++ ftp-server.py
@@ -19,6 +19,17 @@ login <user>
 userdir = os.path.join(os.getcwd(), "docs") #директория по умолчанию
 current_user=""
 
+def change_port(port,server_socket):
+    """Проверка, занят ли порт с этим сокетом. При занятости порта происходит его инкремент."""
+    while(True):
+        try:
+            server_socket.bind(('',port))
+        except socket.error:
+            port+=1
+        else:
+            break
+    return port
+
 def check_access(path):
     """Проверяет, есть ли доступ у пользователя к этой директории """
     fullpath = str() 
@@ -96,76 +107,6 @@ def copysc(filename_from, filename_to):
         filename_to=os.path.join(userdir,filename_from)
     shutil.copyfile(filename_from, filename_to)
     
-def check_user(username,file="users.txt"):
-    """Проверка наличия пользователя в системе по данному IP-адресу"""
-    log_in_successful=bool() 
-    try:
-        user_list=create_user_list(file) 
-    except IOError as e: # если не удалось прочитать из файла, вывести сообщение об ошибке и создать его
-        f=open(file,"w")
-        f.close()
-        print("Файл {} был создан!".format(file))
-        log_in_successful=False
-    finally:
-        user_exists=False
-        for user in user_list: 
-            if username==user[1]: # если пользователь существует, спросить пароль и попробовать залогиниться
-                entered_password=getpass.getpass("Введите пароль: ")
-                while not log_in_user(username,entered_password):
-                    log_in_successful=log_in_user(username, entered_password)
-                user_exists=True
-                break
-        if not(user_exists): # если пользователя не существует, добавить
-            name=input("Введите имя пользователя: ")
-            password=getpass.getpass("Введите пароль: ")
-            add_user(name, password, file)
-            log_in_successful=True
-    if log_in_successful:
-        setup_user(username)
-    return "Пользователь" + username + "вошёл в систему"
-
-def setup_user(user):
-    fullpath=os.path.join(os.getcwd(), user)
-    if not os.path.exists(fullpath):
-        os.mkdir(fullpath)
-    userdir = fullpath
-    current_user=user
-    
-def add_user(name,password,file="users.txt"):
-    """Добавить пользователя с данным именем и паролем"""
-    users_file=open(file,"a") #открывает файл на дозапись
-    name.strip()
-    password.strip() #удаляет лишние пробелы
-    users_file.write("{};{}\n".format(name,encode(password)))
-    print("Пользователь {} добавлен в систему".format(name))
-    users_file.close()
-    
-def create_user_list(file="users.txt"):
-    """Создаёт список с именами и паролями пользователей"""
-    users_file=open(file,"r")
-    user_list=list()
-    for line in users_file: #читает записи из файла, разделяя поля по ; и добавляя в список
-        user=line.split(";")
-        user_list.append(user)
-    return user_list
-    users_file.close()
-            
-def log_in_user(user,entered_password):
-    """Вход пользователя с паролем"""
-    if(encode(entered_password)==user[2].strip()): #шифрование пароля и сравнение с паролем из файла
-        print("Пользователь {} вошёл в систему".format(user[0]))
-        return True
-    else:
-        print("Неверный пароль!")
-        return False
-
-def encode(password):
-    """Безопасное хранение паролей"""
-    return hashlib.sha256(password.encode('utf-8')).hexdigest()
-
-def man(command):
-    pass
-
 def process(req):
     res= ""
     if req == "pwd":
@@ -185,9 +126,7 @@ def process(req):
     elif req.split()[0]=="copysc":
         res = copycs(req.split()[1],req.split()[2])
     elif req.split()[0]=="login":
-        res=check_user(req.split()[1])
-    elif req.split()[0]=="map":
-        res=map(req.split()[1])
+        res=conn.recv(1024).decode()
     else:
         res="bad request"
     return res
@@ -198,15 +137,12 @@ def log(message,file):
     else:
         mod="w+"
         with open(file,mod) as f:
-            f.write(datetime.now() + ": "+ message+"\n")
-
-
-PORT = 6666
+            f.write(str(datetime.now()) + ": "+ message+"\n")
 
 sock = socket.socket()
-sock.bind(("", PORT))
+port = change_port(int(input("Введите номер порта")),sock)
 sock.listen()
-print("Прослушиваем порт ", PORT)
+print("Прослушиваем порт ", port)
 
 while True:
     conn, addr = sock.accept()
@@ -218,4 +154,4 @@ while True:
     conn.send(response.encode())
     log(response,"log.txt")
 
-    conn.close()
+conn.close()
